+++ CHANGELOG.md
@@ -34,6 +34,7 @@ Line wrap the file at 100 chars.                                              Th
 - Change version string parsing to never suggest the user to upgrade to an older version.
 - Make connectivity checker more resilient to suspension.
 - Make uninstaller on desktop platforms attempt to remove WireGuard keys from accounts.
+- Make important notifications not timeout on macOS and remain in the notification list on Linux.
 
 #### Android
 - Show a system notification when the account time will soon run out.